+++ modules/anonify-ecall-types/src/cmd.rs
@@ -1,12 +1,16 @@
-pub const SEND_COMMAND_CMD: u32 = 1;
-pub const FETCH_CIPHERTEXT_CMD: u32 = 2;
-pub const FETCH_HANDSHAKE_CMD: u32 = 3;
+pub const SEND_COMMAND_TREEKEM_CMD: u32 = 1;
+pub const FETCH_CIPHERTEXT_TREEKEM_CMD: u32 = 2;
+pub const FETCH_HANDSHAKE_TREEKEM_CMD: u32 = 3;
 pub const GET_STATE_CMD: u32 = 4;
-pub const JOIN_GROUP_CMD: u32 = 5;
-pub const SEND_HANDSHAKE_CMD: u32 = 6;
+pub const JOIN_GROUP_TREEKEM_CMD: u32 = 5;
+pub const SEND_HANDSHAKE_TREEKEM_CMD: u32 = 6;
 pub const REGISTER_NOTIFICATION_CMD: u32 = 7;
 pub const GET_ENCLAVE_ENCRYPTION_KEY_CMD: u32 = 8;
 pub const SEND_REGISTER_REPORT_CMD: u32 = 9;
 pub const BACKUP_PATH_SECRET_ALL_CMD: u32 = 10;
 pub const RECOVER_PATH_SECRET_ALL_CMD: u32 = 11;
 pub const GET_USER_COUNTER_CMD: u32 = 12;
+
+pub const SEND_COMMAND_ENCLAVE_KEY_CMD: u32 = 13;
+pub const FETCH_CIPHERTEXT_ENCLAVE_KEY_CMD: u32 = 14;
+pub const JOIN_GROUP_ENCLAVE_KEY_CMD: u32 = 15;
